requirements.txt
# Discord
discord.py==2.3.2
py-cord==2.4.1

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
flask-socketio==5.3.4
gunicorn==21.2.0

# Database
pymongo==4.4.1
dnspython==2.4.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
cryptography==41.0.4

# IP Checking
ipaddress==1.0.23
maxminddb==2.4.0

# Async
asyncio==3.4.3
aiohttp==3.8.5

# Development
black==23.7.0
flake8==6.0.0